==> 8. Use Cases/timeseries-missing-values-imputation/4_Results.ipynb <==
# Results Validation
This notebook loads the reconstructed data and compares the results with the available data.

## 0 - Setup

### 0.1 - Imports
Load the necessary dependencies.

%%capture 
!pip install ipywidgets 

%%capture
from ydata.connectors import LocalConnector
from ipywidgets import interact

## 0.2 - Auxiliary Functions
The auxiliary functions are custom-designed utilities developed for the REE use case.

from utils import setting_index_data
from imputation import resample_station_data
from results import meter_statistics
from results_viz import plot_timeseries, compare_acf, plot_multicolor_line, plot_hist, compare_lags, plot_metric_byperiod, polar_graph

## 1 - Load Data

connector = LocalConnector()

# Load the reconstructed data for the given device
reconstructed = setting_index_data(connector.read_file('whole_year_reconstructed.csv'))

# Load the original data for the given device
original = setting_index_data(connector.read_file('train_allmeters.csv'))

## 2 - Results per Meter

@interact(meter=set(reconstructed['station']))
def plot_meter_results(meter):
    meter_reconstructed = reconstructed[reconstructed['station']==meter]
    meter_original = original[original['station']==meter]
    
    # Convert the original data into resampled for full year 
    original_year = resample_station_data(meter_original)

    meter_statistics(original_year)
    plot_timeseries(data=original_year, feature='speed', title='Original Data for '+meter)
    plot_multicolor_line(original_year, meter_reconstructed, title='Reconstruction of '+meter)
    compare_acf(original_year, meter_reconstructed, title='Hourly Autocorrelation')
    plot_hist(original_year, meter_reconstructed, title='Distribution Overlap - Original and Reconstructed Series - '+meter)
    compare_lags(original_year, meter_reconstructed, 'speed')
    plot_metric_byperiod(original_year, reconstructed, 'speed')
    polar_graph(original_year, reconstructed)
